MiniMax M3: fp8 (e4m3) index-K side cache
At decode, the dominant linear-in-context cost for M3 at long context is
the lightning-indexer scan: 57 sparse layers each re-read their full
per-token index-K side cache every step (bf16, 128 B/token/layer -- 5.6 GB
per step at 769K tokens), and the Triton scorer is DRAM-bandwidth-bound.
The scores only feed a top-k block ranking (attention reads the main KV
cache), and index keys are RMS-normed to O(1) magnitudes, so an e4m3 cast
preserves selection while halving the read.

Implements the 'fp8' arm of the existing indexer_kv_dtype config knob:
- side cache allocates as float8_e4m3fn (also halves its KV-pool share)
- the two scorer kernels upcast K loads with .to(q.dtype) -- a no-op for
  bf16 caches, so default behavior is unchanged
- the insert paths already cast via .to(idx_cache.dtype); no writer change

Opt-in via --attention-config '{"indexer_kv_dtype": "fp8"}'.

Measured on 4x RTX PRO 6000 Blackwell (TP4, fp8 main KV, B12X_ATTN, batch 1):
decode 38.8 -> 51.2 tok/s at 384K and 23.5 -> 37.8 tok/s at 769K tokens
(+61%); KV pool at the 1M-token config 1.24M -> 1.43M tokens; short-context
decode unchanged. Needle retrieval verified exact at 751K tokens; fp8-vs-bf16
block-score correlation 1.0000 with top-16 selection agreeing everywhere
above the noise floor (new regression test).

Same design as the fp8 indexer keys DeepSeek V3.2's DSA ships natively.

+++ vllm/models/minimax_m3/common/ops/index_topk.py
@@ -145,7 +145,7 @@ def _index_block_score_kernel(
             + page * stride_ik_blk
             + off_k[None, :] * stride_ik_pos
             + off_d[:, None] * stride_ik_d,
-        )
+        ).to(q.dtype)  # upcast: the index cache may be fp8
         qk = tl.dot(q, k)
         # apply causal mask as needed
         if q_start < i + BLOCK_SIZE_K:
@@ -372,7 +372,7 @@ def _decode_index_score_kernel(
             + page * stride_ik_blk
             + off_k[:, None] * stride_ik_pos
             + off_d * stride_ik_d,
-        )  # [N,D]
+        ).to(q.dtype)  # [N,D] (upcast: the index cache may be fp8)
         kq = tl.dot(k, q)  # [N,HQ]
         kq = tl.where(pos_mask & q_mask[None, :], kq, float("-inf"))
         score = tl.max(kq, axis=0)  # [HQ]
